# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shift API
  description: |
    One API serving both the Shift web app and the future React Native app,
    with zero platform-specific branching in the API layer.

    * Auth uses Bearer access tokens (short-lived JWT) + rotating refresh tokens.
    * All mutating `POST`/`PATCH` requests accept an `Idempotency-Key` header.
    * All timestamps are ISO 8601 strings in UTC.
    * All IDs are UUIDv4 strings.
    * List endpoints that can grow unbounded use cursor-based pagination.
  version: "1.0.0"
  contact:
    name: Shift
servers:
  - url: https://api.shift.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /goals:
    get:
      tags: [Goals]
      summary: List goals
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Array of goals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Goal" }
    post:
      tags: [Goals]
      summary: Create a goal
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, roadmap]
              properties:
                title: { type: string }
                roadmap: { $ref: "#/components/schemas/Roadmap" }
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Goal" }
  /goals/{id}:
    parameters:
      - $ref: "#/components/parameters/GoalId"
    get:
      tags: [Goals]
      summary: Get a goal, including its roadmap
      responses:
        "200":
          description: Full goal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Goal" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Goals]
      summary: Archive/unarchive a goal
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                archived: { type: boolean }
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Goal" }
    delete:
      tags: [Goals]
      summary: Hard delete a goal (rare; prefer archive)
      responses:
        "204": { description: Deleted }
  /goals/{id}/roadmap:
    parameters:
      - $ref: "#/components/parameters/GoalId"
    get:
      tags: [Goals]
      summary: Get milestones + tasks for a goal
      responses:
        "200":
          description: Roadmap
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Roadmap" }

  # ------------------------------------------------------------- PROGRESS
  /goals/{id}/progress:
    parameters:
      - $ref: "#/components/parameters/GoalId"
    post:
      tags: [Progress]
      summary: Log progress and advance the roadmap
      description: |
        Replaces both `completeDebrief` and `logProgressFromChat` from the
        legacy frontend store. `source` is the only difference; the server
        runs the same roadmap-advancement logic regardless of origin
        (mark next task done -> advance milestone -> advance goal -> update streak).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskTitle, rawText, focusMinutes, source]
              properties:
                taskTitle: { type: string }
                rawText: { type: string }
                aiSummary:
                  type: string
                  description: Optional; if omitted and rawText is present, the server may generate it.
                focusMinutes: { type: integer, minimum: 0 }
                source: { type: string, enum: [focus_session, planner_chat] }
      responses:
        "200":
          description: Progress recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  goal: { $ref: "#/components/schemas/Goal" }
                  milestoneAdvanced: { type: boolean }
                  goalCompleted: { type: boolean }
                  streak: { $ref: "#/components/schemas/Streak" }

  # ------------------------------------------------------------ AI PLANNER
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Notification" }
  /notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                read: { type: boolean }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived (15 min) access token. Refresh via POST /auth/refresh.
  parameters:
    GoalId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Client-generated UUID. Server caches the response for 24h and replays it on retry.
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque pagination cursor from a previous response's nextCursor.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 30, minimum: 1, maximum: 100 }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: TASK_NOT_FOUND }
            message: { type: string, example: "No task with that id on this goal's roadmap." }
            requestId: { type: string, example: "req_8f2a..." }
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        nextCursor:
          type: string
          nullable: true
...
    Streak:
      type: object
      properties:
        count: { type: integer }
        longest: { type: integer }
        lastCompletionDay: { type: string, format: date, nullable: true }
...
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        estimateMinutes: { type: integer }
        difficulty: { type: string, enum: [easy, medium, hard] }
        done: { type: boolean }
    Milestone:
      type: object
      properties:
        id: { type: string, format: uuid }
        week: { type: integer }
        title: { type: string }
        status: { type: string, enum: [done, current, upcoming] }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/Task" }
    Roadmap:
      type: object
      properties:
        milestones:
          type: array
          items: { $ref: "#/components/schemas/Milestone" }
    Goal:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        archived: { type: boolean }
        completed: { type: boolean }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        roadmap: { $ref: "#/components/schemas/Roadmap" }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }